pyproject.toml
[project]
name = "mindscape-ama-format"
version = "0.1.0"
description = "Utility to auto-format Mindscape AMA podcast transcripts."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"

dependencies = [
    "openai>=1.93",
    "numpy>=2.3",
    "python-dotenv>=1.1",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
    "ruff>=0.12",
    "baml-py>=0.90",
]
